Fix missing context in Nextjs username/password docs
I found some mistakes(?), I don't know what to call these but the req, res is undefined in the following examples when you try to use it.
If you're like me I usually copy paste docs and this fooled me for a solid 5 minutes, so just contributing so others' time isn't wasted by it.

Great library btw. Easy to setup and it just works.